Index Page Update
This is the update for the homework due 4/7. The links are there and I've applied some styling. Still having some trouble with alignment and having certain elements do what I want them to. I added some comments in the code to try to explain my thinking and to show what I've tried.